mini_cursor/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-dotenv==1.0.1


requests==2.32.3


openai>=1.35.13
langfuse==2.37.0


typing-extensions>=4.12.2
streamlit==1.37.1


